pish-examples/experiment-client/requirements.txt
python-vlc==3.0.7110
requests==2.18.1

##### Core scientific packages
jupyter==1.0.0
matplotlib==3.1.2
numpy==1.18.0
pandas==0.25.3
scipy==1.4.1

# Visualize
seaborn==0.9.0
pydot==1.4.1
